Houston we have parse
Hacked together proof of concept level code... we now download + parse
the rss from all subs, and update the titles of the podcasts to the
titles given in the rss feeds.

Absolutely not architected properly, and the service code is super ugly,
but i'm tired and done for the night...just wanted to see "first light".